Fix sending future data to inference server
Humans with time_slot at hour 0 don't have the time to update their data before it gets sent to the inference server. 

Previous fix was wrong by waiting a day to send those data. 

This code is a better fixes it by sending a day old data instead of waiting a day before sending the data

Proper fix would be to update the data at the same time it gets sent to the inference server but that would require a significant refactor of the Human.run() if we want to put it there